Make some tests use mocks
